+++ tests/views/test_customer_billing_info_view.py
@@ -10,12 +10,13 @@ class TestCustomerBillingInfo(BaseTestView):
     """Test Billing Info for an Asset."""
 
     base_path = (
-        '/customers/c2034c1b-0a40-4b84-9ace-54b958f64ed4/billing'
+        '/billing_info/customers'
     )
     dependencies = [
         (models.Customer, 'data/customers.json'),
     ]
     file_path = 'data/customer_billing_infos.json'
+    ignore_validation_fields = ['customer', 'state_history']
     model = models.CustomerBillingInfo
     UPDATE_SUCCESS_MESSAGE = ''
     NOT_FOUND_MESSAGE = ''